refactor : 다리 생성 로직을 관리하는 패키지명을 point -> bridge로 수정
