+++ src/test/java/domain/bridge/strategy/BridgeGeneratorStub.java
@@ -1,7 +1,7 @@
-package domain.point.strategy;
+package domain.bridge.strategy;
 
 import domain.LadderBridge;
-import domain.point.BridgeGenerator;
+import domain.bridge.BridgeGenerator;
 
 import java.util.List;
